+++ lib/generators/settings/templates/model.rb
@@ -1,3 +1,5 @@
 # RailsSettings Model
 class <%= class_name %> < RailsSettings::CachedSettings
+  source Rails.root.join("config/app.yml")
+  namespace Rails.env
 end