+++ lib/rails-settings/yml_setting.rb
@@ -1,0 +1,6 @@
+require 'settingslogic'
+
+module RailsSettings
+  class YMLSetting < Settingslogic
+  end
+end