+++ spec/rails-settings-cached/scoped_setting_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
 
 describe RailsSettings::ScopedSettings do
-  it "extends `CachedSettings`" do
-    expect(described_class.ancestors).to include(RailsSettings::CachedSettings)
+  it "extends `RailsSettings::Base`" do
+    expect(described_class.ancestors).to include(RailsSettings::Base)
   end
 end